[CON-49] Further clean up
