+++ lib/presentation/home_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import '../domain/service/verb_manager.dart';
+import '../domain/service/verb_service.dart';
 import 'grammar/grammar_screen.dart';
 import 'quiz/screens/quiz_screen.dart';
 import 'quiz/view_models/quiz_view_model.dart';
@@ -23,12 +23,12 @@ class _HomeScreenState extends State<HomeScreen> {
   final List<Widget> _screens = [
     GrammarScreen(),
     // Quiz Screen
-    ChangeNotifierProxyProvider<VerbManager, QuizViewModel>(
+    ChangeNotifierProxyProvider<VerbService, QuizViewModel>(
       create: (_) => QuizViewModel(),
       update: (_, verbManager, quizViewModel) => quizViewModel!..updateVerbs(verbManager.verbs),
       child: QuizScreen(),
     ),
-    ChangeNotifierProxyProvider<VerbManager, SearchViewModel>(
+    ChangeNotifierProxyProvider<VerbService, SearchViewModel>(
       create: (_) => SearchViewModel(),
       update: (_, verbManager, searchViewModel) => searchViewModel!..updateVerbs(verbManager.verbs),
       child: VerbsScreen(),
