+++ lib/presentation/settings/screens/settings_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import '../../../domain/service/verb_manager.dart';
+import '../../../domain/service/verb_service.dart';
 import '../../../utilities/app_values.dart';
 import '../../../utilities/extensions/build_context_extensions.dart';
 import '../widgets/choose_pronouns_sheet.dart';
@@ -28,7 +28,7 @@ class SettingsScreen extends StatelessWidget {
       context: context,
       builder: (_) {
         return Provider.value(
-          value: context.read<VerbManager>(),
+          value: context.read<VerbService>(),
           child: SingleChildScrollView(
             scrollDirection: Axis.vertical,
             child: ChooseVerbsSheet(),
@@ -43,7 +43,7 @@ class SettingsScreen extends StatelessWidget {
       context: context,
       builder: (_) {
         return Provider.value(
-          value: context.read<VerbManager>(),
+          value: context.read<VerbService>(),
           child: SingleChildScrollView(
             scrollDirection: Axis.vertical,
             child: ChoosePronounsSheet(),
