+++ lib/presentation/verbs/screens/verbs_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import '../../../domain/service/verb_manager.dart';
+import '../../../domain/service/verb_service.dart';
 import 'verbs_content.dart';
 
 final _logTag = (VerbsScreen).toString();
@@ -17,13 +17,13 @@ class VerbsScreen extends StatelessWidget {
     return Scaffold(
       appBar: AppBar(title: Text('Verbs 📚')),
       body: FutureBuilder(
-        future: context.read<VerbManager>().initializationFuture,
+        future: context.read<VerbService>().initializationFuture,
         builder: (context, snapshot) {
           if (snapshot.connectionState == ConnectionState.waiting) {
             return Center(child: CircularProgressIndicator());
           } else if (snapshot.hasError) {
             return Center(child: Text('Error: ${snapshot.error}'));
-          } else if (context.read<VerbManager>().verbs.isEmpty) {
+          } else if (context.read<VerbService>().verbs.isEmpty) {
             return Center(child: Text('No verbs available'));
           } else {
             return VerbsContent(
